"던전 입장 및 던전 시작 로직구현"
Feat : "던전 입장 및 던전 시작 로직구현"

Body : 던전 입장 및 던전 시작 구현 완료, 던전 플레이 도중 버그 발견으로 버그 수정중

Footer :

+++ TRPG/Item.cs
@@ -1,10 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Threading.Tasks;
+
 
 namespace TRPG
 {
@@ -13,16 +8,16 @@ namespace TRPG
         public int Id { get; set; }  // 아이템을 판매 구매 방식에 있어 index오류로 인해
                                      // 고유한 ID를 부여하여 판매하는 방식
         public string Name {  get; set; }
-        public float Armor {  get; set; }
-        public float Attack { get; set; }
-        public float Health { get; set; }
+        public int Armor {  get; set; }
+        public int Attack { get; set; }
+        public int Health { get; set; }
         public float Price { get; set; }
         public float SellPrice { get; set; } 
         public string ItemInfo {  get; set; }
         public bool isPurchased {  get; set; }
         public bool isEquipped { get; set; }
 
-        public Item(string name, float attack, float armor, float health, string itemInfo, float price) 
+        public Item(string name, int attack, int armor, int health, string itemInfo, float price) 
         {
             Name = name;
             Armor = armor;
